Bug 1579862 - Tweaking HuffmanTableImplementationSaturated;r=arai
A few changes to HuffmanTableImplementationSaturated:

- we can now create instances of HuffmanTableImplementationSaturated with max bit lengths up to 10;
- we reject cases in which we have more than 256 elements;
- internal indices in HuffmanTableImplementationSaturated are now represented with uint8_t instead of usize_t,
  which divides the size of this array by 4 and should improve memory locality.

According to my benchmarking, this decreases duration by ~7%.

Depends on D45644

Differential Revision: https://phabricator.services.mozilla.com/D45722

UltraBlame original commit: ebb8ad9cd82eae549eb6e96395fdde61f48fb78a

+++ js/src/frontend/BinASTTokenReaderContext.h
@@ -465,8 +465,7 @@ class HuffmanTableImplementationSaturated {
   
   
   
-  
-  Vector<size_t> saturated;
+  Vector<uint8_t> saturated;
